+++ aidge-20260428/ChangeLog.md
@@ -1,3 +1,10 @@
+2026-06-29 Version: 2.0.0
+- Support API ImageTranslationPlus.
+- Support API MaterialInspection.
+- Update API ImageTranslationPro: add request parameters Async.
+- Update API ImageTranslationPro: delete request parameters callType.
+
+
 2026-06-16 Version: 1.0.0
 - Generated python 2026-04-28 for Aidge.